+++ packages/react/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @tdv/react
+# @tdeepv/react
+
+## 3.0.0
+
+### Major Changes
+
+- Add new white color
 
 ## 2.0.0