+++ packages/tokens/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @tdv/tokens
+# @tdeepv/tokens
+
+## 3.0.0
+
+### Major Changes
+
+- Add new white color
 
 ## 2.0.0