+++ packages/ts-config/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @tdv/ts-config
+# @tdeepv/ts-config
+
+## 2.1.0
+
+### Minor Changes
+
+- Add new white color
 
 ## 2.0.0